+++ src/main/java/com/example/VeckorVorker/dao/WorkerRepository.java
@@ -3,5 +3,5 @@ package com.example.VeckorVorker.dao;
 import com.example.VeckorVorker.model.Worker;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface  WorkerRepository extends JpaRepository<Worker, Long> {
+public interface  WorkerRepository extends JpaRepository<Worker, Integer> {
 }
